comm: pho_comm_open: allow binding interface for tcp server
It can be useful to bind a listening socket to an interface, so 0.0.0.0
can be used in order to allow connections on an IP address that is not yet
configured (e.g. virtual IP), but still limit inbound connections from a
single interface without filtering at firewall level.

This adds a new field to struct pho_comm_addr's tcp struct so that if an
interface is set we use it.

setsockopt SO_BINDTODEVICE requires NET_ADMIN so this call might fail:
this patch ignores the error on purpose so the server still starts,
but a log is emitted for diagnostic.

Change-Id: I5c49758a96213ba6882ff1a85ad82e72abf64c8c