admin: handle `lost` parameter for media delete
This patch adds the handling of the `lost` parameter of the media delete
function. If true, we will not skip over media containing extents, and
instead of simply using the `dss_media_delete` function, call a new
one to remove the media, the extents marked as on it, and update the
objects and copies as necessary.

Change-Id: Ibe996099d441cb38929646606d496d5aaafb2546
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>