grouped_write: add request data structures
Write requests will be stored in queues in a hash table.
The queues will be hashed based on the tags and layout
parameters of the write requests. These queues will be
divided into one subqueue per grouping. Groupings will
be written one after the other to increase the data
locality on the tape.

Change-Id: I0db73965f106e751d5e33fda70bd6e3ba6ca463e