grouped_write: implement .push_request method
If a queue has already been allocated for this kind of request,
we push it on the queue. Otherwise, we create a new queue.

When pushing to a queue, we try to find the corresponding grouping.
If not found, we create a new one. There is a special case for if
there is no grouping that needs to be managed since the grouping's
name is the key used in the queue's grouping_index hashtable.

Change-Id: Ibd12c52000a19c93c9470089a122158aba0a8b5c