grouped_write: implementation of .peek_request
This function will simply try to find from which queue
we should pick a request and return that request.

If some devices are free, try to allocate the next queue by moving it to
the allocated list. Then find the first queue whose devices are free to
do an I/O and return the first element of this queue. At this point, the
queue may not have devices associated to it yet.

Change-Id: I2feb13eebd28031fb52e6d1bf517a7b18330acaf