grouped_write: handle .get_device_medium_pair retries
get_device_medium_pair can be called multiple times with the
same index as long as the upper layer is not satisfied with
the device that was returned. A device can for instance become
failed asynchronously.

get_device_medium_pair can also be called after a device has
already been allocated to the queue for a previous request.

We need to manage both cases differently.

When reallocating on the same index, we simply need to remove
the current device from the queue and try to find a new one.

When allocating a request on a queue that was already initialized,
we don't need to do much. We simply recheck the medium's write
compatibility since an admin can for instance lock a medium while
some I/O are scheduled on it.

Since in both cases, there is already a device in the queue, we need
to know whether we are trying to reallocate or not. This is done
by remembering the last index used when calling get_device_medium_pair.

Change-Id: Ia655294db62eddfad285b0a1fec1010919efda5b