lost: add test for raid1 and the skeleton for raid4
This patch add a test for lost media with raid1 objects on it.
It checks that objects/copies/media/extents are correctly
deleted/updated.

This patch also adds the skeleton for a raid4 test, however since the
raid4 layout doesn't implement the `reconstruct` function, it cannot be
fully tested yet.

Change-Id: I646fc3c0947ef7bcfd2b30a8846b0a50a58626b8
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>