lrs: add last_locate check before picking a device
In case a device needs to be picked up, we check its medium is not
reserved for a future `get` operation before switching it to the
one requested by the client.

A test is added to check the daemon has the right behavior i.e.
fail a request if a drive/medium is reserved for a locate.

Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud.externe@ayudata.fr>