tests: add phobosd-client program
This program will allow writing custom scripts to
send specific requests to phobosd. This makes it
possible to write tests that would be otherwise
difficult to do using regular phobos commands.

This program is a simple interpreter that generates
and sends requests to phobosd. For instance:

```
# send a write alloc with 2 media of family tape with tags foo and bar
# The request is sent with the grouping baz
write -t foo,bar -g baz -f tape 2

# wait for a write alloc response for request id 1 (i.e. the write above)
recv 1 write_alloc

# release the media allocated for the first write
release 1

# wait for the write alloc release response
# the release's request ID is 2
recv 2 write_release
```

Request IDs are incremented for each new request. read, write and release
requests will increase the current ID. Therefore, when expecting a
response to a specific request, the ID can be provided as an argument
to the release or recv commands.

Change-Id: Iae10a104b8406a94e87cb4c066ea14e745d47a77