lrs: prevent negative physical space free update
Repack, with its huge amount of newly written extents, could try to update a
negative physical space free at release. This is due to the fact that we ignore
automatic compression that happens on magnetic tape. The correct value will be
set by the statfs at sync. Meanwhile, we set the secure 0 value.

Negative value is not accepted by dss_media_get.

This patch fix the github issue #53 .