lost: improve reconstruction logic for raid4
This patch improves the reconstruction logic for raid4 by using the
cyclical layout index of each extent to determine the size of each part
of a copy. This allows us to better determine if an object is still
readable or not, as we rely on the actual size of each part instead of
offsets.

Change-Id: Ic2e9640bc5c51fe8df2d382c450fa14fe08c89ef
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>