layout: retrieve object size on decode
This patch changes the decode logic to use the `object_size` added as
part of the layout information rather than calculate it with the extents
obtained from the database.

Since those extents can now be removed and the copies updated, a copy
can still be readable even though the number of extents is not a
multiple of the replica count for raid1 or 3 for raid4. Therefore,
relying on this extent count is now invalid.

In the same time, the decoder would use the extents retrieved from the
database and their size to determine the size of the object to read.
Since that information has now been added to the layout information,
there is no need calculate it on the fly.

Change-Id: If28a0b4f2c41b0052bc86bf5e0c6e5af8d8a1f21
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>