phobosd-client: keep initial write request for alloc size
phobosd returns the maximum available size in all the media returned
in the write alloc. We need to set the size initially asked in
the write request when making the corresponding release request.
Add the write allocation requests in a hash table so that we
can retrieve them when making the release request.

Change-Id: Idcec014823e67ea0c758a35e9d1933fb00776301