locate: change asserts to proper checks
This patch changes the asserts done during a `phobos_locate` to proper
checks. It is needed because there are cases where the DSS could
contain an object but no associated copy or layout. This may occur for
instance if a `put` process is killed before the end, leaving an
incomplete object or copy in the database, which would cause assert
failures during the `locate` as we assume there is a valid copy and
layout for each object registered.

Change-Id: I9141aa7885f1c6279e2fcbffe05af5c301b4c723
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>